# new_project/contracts/openapi/investment-api.yaml
openapi: 3.1.0
info:
  title: Personal Investment API
  version: 0.2.0
  description: 本地开发切片。所有金额以原币种两位小数最小单位的十进制字符串传输，绝不使用浮点数。
servers:
  - url: http://127.0.0.1:<API_PORT>
paths:
...
  /api/v1/strategies:
    get:
      summary: 读取四个策略工作区摘要
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 即使账本为空仍返回四张策略卡片
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyListResponse'
        '401':
          $ref: '#/components/responses/Problem'
...
  /api/v1/strategies/scans:
    post:
      summary: 异步扫描已存策略输入
      description: 仅读取已落库的账本、市场快照和参考净值，不触发外部行情请求或交易。
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyScanRequest'
      responses:
        '202':
          description: 策略扫描已持久化入队
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyScan'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/strategies/scans/{strategyScanId}:
    get:
      summary: 查询策略扫描及逐策略结果
      security:
        - bearerAuth: []
      parameters:
        - name: strategyScanId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
      responses:
        '200':
          description: 扫描状态及已追加的逐策略结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyScan'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /api/v1/strategies/{strategyKey}/workspace:
    get:
      summary: 读取一个策略工作区及可用账务动作
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
      responses:
        '200':
          description: 当前策略工作区摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyWorkspaceResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
  /api/v1/strategies/{strategyKey}/rule-versions:
    get:
      summary: 分页读取策略规则版本历史
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: 按创建时间和业务 ID 倒序的稳定游标页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleVersionPage'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
    post:
      summary: 追加并激活一个策略规则版本
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyRuleVersionCreate'
      responses:
        '201':
          description: 新版本已成为唯一活动规则，旧活动版本已归档
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyRuleVersionCreated'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/strategies/{strategyKey}/reference-nav:
    get:
      summary: 分页读取 USD 策略参考净值事实
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: 按创建时间和业务 ID 倒序的稳定游标页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceNavPage'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
    post:
      summary: 追加用户明确维护的 USD 策略参考净值
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferenceNavCreate'
      responses:
        '201':
          description: 已追加参考净值事实
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceNav'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /api/v1/strategies/{strategyKey}/evaluations:
    get:
      summary: 分页读取不可变策略评估快照
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 30
      responses:
        '200':
          description: 按评估时点和业务 ID 倒序的稳定游标页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyEvaluationPage'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
    post:
      summary: 基于已存账本与市场快照生成评估
      security:
        - bearerAuth: []
      parameters:
        - name: strategyKey
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/StrategyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StrategyEvaluationRequest'
      responses:
        '200':
          description: 已返回或追加评估快照；DATA_STALE 和 BLOCKED 仍为正常解释性结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategyEvaluation'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
    Problem:
      description: RFC 7807 风格错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    StrategyKey:
      type: string
      enum: [HIGH_DIVIDEND, QQQ_GROWTH, IC_IM, DEEP_PUT]
    StrategyCard:
      type: object
      required: [strategyKey, displayName, currency, status, message]
      properties:
        strategyKey:
          $ref: '#/components/schemas/StrategyKey'
        displayName:
          type: string
        currency:
          type: string
          enum: [CNY, USD]
        activeRuleVersionId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        inputAt:
          type: [string, 'null']
          format: date-time
        status:
          type: string
          enum: [IN_RANGE, WATCH, BLOCKED, DATA_STALE, CROSS_CURRENCY_UNVALUED]
        message:
          type: string
    StrategyListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 4
          maxItems: 4
          items:
            $ref: '#/components/schemas/StrategyCard'
    StrategyWorkspaceResponse:
      type: object
      required: [strategy, availableActions, activeRule, latestEvaluation]
      properties:
        strategy:
          $ref: '#/components/schemas/StrategyCard'
        availableActions:
          type: array
          items:
            type: string
        activeRule:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/StrategyRuleVersionHistory'
        latestEvaluation:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/StrategyEvaluation'
    StrategyRuleVersionCreate:
      type: object
      required: [ruleVersion, rule]
      properties:
        ruleVersion:
          type: string
          minLength: 1
          maxLength: 64
        expectedActiveRuleVersionId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        rule:
          type: object
          description: 所有金额必须是 *_cent 十进制字符串并带显式同前缀 currency 字段。
          additionalProperties: true
    StrategyRuleVersionCreated:
      type: object
      required: [strategyRuleVersionId, strategyKey, ruleVersion, status, createdAt]
      properties:
        strategyRuleVersionId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        strategyKey:
          $ref: '#/components/schemas/StrategyKey'
        ruleVersion:
          type: string
        status:
          type: string
          enum: [ACTIVE]
        createdAt:
          type: string
          format: date-time
    StrategyRuleVersionHistory:
      type: object
      required: [strategyRuleVersionId, strategyKey, ruleVersion, rule, status, createdAt]
      properties:
        strategyRuleVersionId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        strategyKey:
          $ref: '#/components/schemas/StrategyKey'
        ruleVersion:
          type: string
        rule:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        createdAt:
          type: string
          format: date-time
    RuleVersionPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StrategyRuleVersionHistory'
        nextCursor:
          type: [string, 'null']
    ReferenceNavCreate:
      type: object
      required: [referenceNavCent, currency, asOfAt, validUntil]
      properties:
        referenceNavCent:
          type: string
          pattern: '^[1-9][0-9]*$'
        currency:
          type: string
          enum: [USD]
        asOfAt:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
    ReferenceNav:
      type: object
      required: [strategyReferenceNavId, strategyKey, currency, referenceNavCent, asOfAt, validUntil, source]
      properties:
        strategyReferenceNavId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        strategyKey:
          $ref: '#/components/schemas/StrategyKey'
        currency:
          type: string
          enum: [USD]
        referenceNavCent:
          type: string
          pattern: '^[1-9][0-9]*$'
        asOfAt:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        source:
          type: string
          enum: [MANUAL]
    ReferenceNavPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReferenceNav'
        nextCursor:
          type: [string, 'null']
    StrategyEvaluationRequest:
      type: object
      properties:
        asOfDate:
          type: string
          format: date
    StrategyEvaluation:
      type: object
      required: [strategyEvaluationId, strategyKey, inputVersion, asOfAt, status, result]
      properties:
        strategyEvaluationId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        strategyKey:
          $ref: '#/components/schemas/StrategyKey'
        strategyRuleVersionId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        inputVersion:
          type: string
        asOfAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [IN_RANGE, WATCH, BLOCKED, DATA_STALE, CROSS_CURRENCY_UNVALUED]
        result:
          type: object
          additionalProperties: true
    StrategyEvaluationPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StrategyEvaluation'
        nextCursor:
          type: [string, 'null']
...
    StrategyScanRequest:
      type: object
      properties:
        strategyKeys:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/StrategyKey'
    StrategyScan:
      type: object
      required: [strategyScanId, asOfAt, status, attemptNo, items]
      properties:
        strategyScanId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        asOfAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [QUEUED, RUNNING, SUCCEEDED, PARTIAL_SUCCEEDED, FAILED, SKIPPED_STALE]
        attemptNo:
          type: integer
          minimum: 0
        startedAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        result:
          type: [object, 'null']
          additionalProperties: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/StrategyScanItem'
    StrategyScanItem:
      type: object
      required: [strategyScanItemId, strategyKey, status, createdAt]
      properties:
        strategyScanItemId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        strategyKey:
          $ref: '#/components/schemas/StrategyKey'
        strategyEvaluationId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        status:
          type: string
          enum: [IN_RANGE, WATCH, BLOCKED, DATA_STALE, CROSS_CURRENCY_UNVALUED, FAILED]
        failureCode:
          type: [string, 'null']
        failureMessage:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...
    Problem:
      type: object
      required: [code, message, traceId, details]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          type: array
          items:
            type: string